requirements.txt
tensorflow==2.20.0
tensorflow-serving-api==2.19.1
fastapi[standard]
uvicorn
python-multipart
pillow
matplotlib
numpy